+++ lib/hmmlearn/hmm.py
@@ -8,6 +8,7 @@ import logging
 
 import numpy as np
 from scipy import linalg, special
+from scipy.stats import poisson
 from sklearn import cluster
 from sklearn.utils import check_random_state
 
@@ -16,14 +17,14 @@ from .stats import log_multivariate_normal_density
 from .base import BaseHMM
 from .utils import fill_covars, log_mask_zero, log_normalize, normalize
 
-__all__ = ["GMMHMM", "GaussianHMM", "MultinomialHMM"]
+__all__ = ["GMMHMM", "GaussianHMM", "MultinomialHMM", "PoissonHMM"]
 
 
 _log = logging.getLogger(__name__)
 COVARIANCE_TYPES = frozenset(("spherical", "diag", "full", "tied"))
 
 
-def _check_and_set_gaussian_n_features(model, X):
+def _check_and_set_n_features(model, X):
     _, n_features = X.shape
     if hasattr(model, "n_features"):
         if model.n_features != n_features:
@@ -193,7 +194,7 @@ class GaussianHMM(BaseHMM):
         }
 
     def _init(self, X):
-        _check_and_set_gaussian_n_features(self, X)
+        _check_and_set_n_features(self, X)
         super()._init(X)
 
         if self._needs_init("m", "means_"):
@@ -661,7 +662,7 @@ class GMMHMM(BaseHMM):
         }
 
     def _init(self, X):
-        _check_and_set_gaussian_n_features(self, X)
+        _check_and_set_n_features(self, X)
         super()._init(X)
         nc = self.n_components
         nf = self.n_features
@@ -1049,3 +1050,170 @@ class GMMHMM(BaseHMM):
                 c_d = (stats['post_sum'] + nm + nus + nf + 1)[:, None, None]
 
             self.covars_ = c_n / c_d
+
+
+class PoissonHMM(BaseHMM):
+    """
+    Hidden Markov Model with Poisson emissions.
+
+    Attributes
+    ----------
+    monitor_ : ConvergenceMonitor
+        Monitor object used to check the convergence of EM.
+
+    startprob_ : array, shape (n_components, )
+        Initial state occupation distribution.
+
+    transmat_ : array, shape (n_components, n_components)
+        Matrix of transition probabilities between states.
+
+    lambdas_ : array, shape (n_components, n_features)
+        The expectation value of the waiting time parameters for each
+        feature in a given state.
+    """
+
+    def __init__(self, n_components=1, startprob_prior=1.0,
+                 transmat_prior=1.0, lambdas_prior=None,
+                 lambdas_weight=None,
+                 algorithm="viterbi", random_state=None,
+                 n_iter=10, tol=1e-2, verbose=False,
+                 params="stl", init_params="stl",
+                 implementation="log"):
+        """
+        Parameters
+        ----------
+        n_components : int
+            Number of states.
+
+        startprob_prior : array, shape (n_components, ), optional
+            Parameters of the Dirichlet prior distribution for
+            :attr:`startprob_`.
+
+        transmat_prior : array, shape (n_components, n_components), optional
+            Parameters of the Dirichlet prior distribution for each row
+            of the transition probabilities :attr:`transmat_`.
+
+        lambdas_prior, lambdas_weight : array, shape (n_components,), optional
+            The gamma prior on the lambda values using alpha-beta notation,
+            respectivley. If None, will be set based on the method of
+            moments.
+
+        algorithm : {"viterbi", "map"}, optional
+            Decoder algorithm.
+
+        random_state: RandomState or an int seed, optional
+            A random number generator instance.
+
+        n_iter : int, optional
+            Maximum number of iterations to perform.
+
+        tol : float, optional
+            Convergence threshold. EM will stop if the gain in log-likelihood
+            is below this value.
+
+        verbose : bool, optional
+            Whether per-iteration convergence reports are printed to
+            :data:`sys.stderr`.  Convergence can also be diagnosed using the
+            :attr:`monitor_` attribute.
+
+        params, init_params : string, optional
+            The parameters that get updated during (``params``) or initialized
+            before (``init_params``) the training.  Can contain any
+            combination of 's' for startprob, 't' for transmat, and 'l' for
+            lambdas.  Defaults to all parameters.
+
+        implementation: string, optional
+            Determines if the forward-backward algorithm is implemented with
+            logarithms ("log"), or using scaling ("scaling").  The default is
+            to use logarithms for backwards compatability.
+        """
+        BaseHMM.__init__(self, n_components,
+                         startprob_prior=startprob_prior,
+                         transmat_prior=transmat_prior,
+                         algorithm=algorithm,
+                         random_state=random_state,
+                         n_iter=n_iter, tol=tol, verbose=verbose,
+                         params=params, init_params=init_params,
+                         implementation=implementation)
+        self.lambdas_prior = lambdas_prior
+        self.lambdas_weight = lambdas_weight
+
+    def _init(self, X):
+        _check_and_set_n_features(self, X)
+        super()._init(X)
+        self.random_state = check_random_state(self.random_state)
+
+        mean_X = X.mean()
+        var_X = X.var()
+
+        if self._needs_init('l', 'lambdas_'):
+            # initialize with method of moments based on X
+            self.lambdas_ = self.random_state.gamma(
+                shape=mean_X**2 / var_X,
+                scale=var_X / mean_X,  # numpy uses theta = 1 / beta
+                size=(self.n_components, self.n_features))
+
+        if self.lambdas_prior is None:
+            self.lambdas_prior = mean_X**2 / var_X
+        if self.lambdas_weight is None:
+            self.lambdas_weight = mean_X / var_X  # use beta notation here
+
+    def _get_n_fit_scalars_per_param(self):
+        nc = self.n_components
+        nf = self.n_features
+        return {
+            "s": nc - 1,
+            "t": nc * (nc - 1),
+            "l": nc * nf,
+        }
+
+    def _check(self):
+        super()._check()
+
+        self.lambdas_ = np.atleast_2d(self.lambdas_)
+        n_features = getattr(self, "n_features", self.lambdas_.shape[1])
+        if self.lambdas_.shape != (self.n_components, n_features):
+            raise ValueError(
+                "lambdas_ must have shape (n_components, n_features)")
+        self.n_features = n_features
+
+    def _generate_sample_from_state(self, state, random_state=None):
+        random_state = check_random_state(random_state)
+        return random_state.poisson(self.lambdas_[state])
+
+    def _compute_log_likelihood(self, X):
+        return np.array([np.sum(poisson.logpmf(X, lambdas), axis=1)
+                         for lambdas in self.lambdas_]).T
+
+    def _compute_likelihood(self, X):
+        return np.array([np.prod(poisson.pmf(X, lambdas), axis=1)
+                         for lambdas in self.lambdas_]).T
+
+    def _initialize_sufficient_statistics(self):
+        stats = super()._initialize_sufficient_statistics()
+        stats['nsamples'] = 0
+        stats['post'] = np.zeros(self.n_components)
+        stats['obs'] = np.zeros((self.n_components, self.n_features))
+        return stats
+
+    def _accumulate_sufficient_statistics(self, stats, obs, lattice,
+                                          posteriors, fwdlattice, bwdlattice):
+        super()._accumulate_sufficient_statistics(
+            stats, obs, lattice, posteriors, fwdlattice, bwdlattice)
+        if 'l' in self.params:
+            stats['nsamples'] += obs.shape[0]
+            stats['post'] += posteriors.sum(axis=0)
+            stats['obs'] += np.dot(posteriors.T, obs)
+
+    def _do_mstep(self, stats):
+        super()._do_mstep(stats)
+
+        if 'l' in self.params:
+            # Based on: Hyvönen & Tolonen, "Bayesian Inference 2019"
+            # section 3.2
+            # https://vioshyvo.github.io/Bayesian_inference
+            alphas, betas = self.lambdas_prior, self.lambdas_weight
+            n = stats['nsamples']
+            kappas = betas / (betas + n)
+            y_bar = stats['obs'] / stats['post'][:, None]
+            self.lambdas_ = kappas * (alphas / betas) + (1 - kappas) * y_bar